requirements.txt
fake-useragent==0.1.11
Flask==1.1.1
jsonpath==0.82
lxml==4.3.4
pymongo==3.8.0
PyMySQL==0.7.11
pyquery==1.4.0
pywin32==223
redis==3.3.6
requests==2.22.0
retrying==1.3.3
Scrapy==1.7.3
scrapy-redis==0.6.8
scrapy-splash==0.7.2
SQLAlchemy==1.3.6